requirements.txt
asgiref==3.5.2
Django==4.0
django-admin-interface==0.19.2
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-froala-editor==4.0.14
django-js-asset==2.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
xlwt==1.3.0
